Move hostname/data to comp, expose hostname interface
