+++ comp/core/hostname/component.go
@@ -7,19 +7,10 @@
 package hostname
 
 import (
-	"context"
-
-	"github.com/DataDog/datadog-agent/pkg/util/hostname/data"
+	"github.com/DataDog/datadog-agent/comp/core/hostname/hostnameinterface"
 )
 
 // team: agent-shared-components
 
 // Component is the component type.
-type Component interface {
-	// Get returns the host name for the agent.
-	Get(context.Context) (string, error)
-	// GetWithProvider returns the hostname for the Agent and the provider that was use to retrieve it.
-	GetWithProvider(ctx context.Context) (data.Data, error)
-	// GetSafe is Get(), but it returns 'unknown host' if anything goes wrong.
-	GetSafe(context.Context) string
-}
+type Component = hostnameinterface.HostnameInterface
