+++ comp/core/hostname/hostnameinterface/data.go
@@ -3,8 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package data provides utilities for hostname and the hostname provider
-package data
+package hostnameinterface
 
 const (
 	ConfigProvider  = "configuration"
