+++ comp/core/hostname/hostnameinterface/go.mod
@@ -1,0 +1,3 @@
+module github.com/DataDog/datadog-agent/comp/core/hostname/hostnameinterface
+
+go 1.21
