--- pkg/util/hostname/data/go.mod
@@ -1,3 +1,0 @@
-module github.com/DataDog/datadog-agent/pkg/util/hostname/data
-
-go 1.21
